backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.2
google-api-python-client==2.108.0
google-auth==2.23.4
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
urllib3==2.0.7
pymed==0.8.9
pydantic==2.5.0